+++ vtm-mvt/src/org/oscim/tiling/source/mvt/MapzenMvtTileSource.java
@@ -1,6 +1,6 @@
 /*
  * Copyright 2013 Hannes Janetzek
- * Copyright 2016-2017 devemux86
+ * Copyright 2016-2018 devemux86
  *
  * This file is part of the OpenScienceMap project (http://www.opensciencemap.org).
  *
@@ -18,6 +18,7 @@
 package org.oscim.tiling.source.mvt;
 
 import org.oscim.tiling.ITileDataSource;
+import org.oscim.tiling.OverzoomTileDataSource;
 import org.oscim.tiling.source.UrlTileDataSource;
 import org.oscim.tiling.source.UrlTileSource;
 
@@ -30,8 +31,9 @@ public class MapzenMvtTileSource extends UrlTileSource {
         private String locale = "";
 
         public Builder() {
-            super(DEFAULT_URL, DEFAULT_PATH, 1, 17);
+            super(DEFAULT_URL, DEFAULT_PATH);
             keyName("api_key");
+            overZoom(16);
         }
 
         public T locale(String locale) {
@@ -66,6 +68,6 @@ public class MapzenMvtTileSource extends UrlTileSource {
 
     @Override
     public ITileDataSource getDataSource() {
-        return new UrlTileDataSource(this, new MvtTileDecoder(locale), getHttpEngine());
+        return new OverzoomTileDataSource(new UrlTileDataSource(this, new MvtTileDecoder(locale), getHttpEngine()), mOverZoom);
     }
 }