+++ vtm/src/org/oscim/layers/tile/vector/labeling/LabelTileLoaderHook.java
@@ -1,6 +1,6 @@
 /*
  * Copyright 2013 Hannes Janetzek
- * Copyright 2016-2017 devemux86
+ * Copyright 2016-2018 devemux86
  * Copyright 2016 Andrey Novikov
  *
  * This file is part of the OpenScienceMap project (http://www.opensciencemap.org).
@@ -123,9 +123,22 @@ public class LabelTileLoaderHook implements TileLoaderThemeHook {
 
             LabelTileData ld = get(tile);
             if (element.type == LINE) {
-                // TODO
+                int offset = 0;
+                for (int i = 0, n = element.index.length; i < n; i++) {
+                    int length = element.index[i];
+                    if (length < 4)
+                        break;
+
+                    WayDecorator.renderSymbol(null, element.points, symbol,
+                            offset, length, ld);
+                    offset += length;
+                }
             } else if (element.type == POLY) {
                 PointF centroid = element.labelPosition;
+                if (!Parameters.POLY_SYMBOL) {
+                    if (centroid == null)
+                        return false;
+                }
                 // skip unnecessary calculations if centroid is outside of visible area
                 if (centroid != null && (centroid.x < 0 || centroid.x > Tile.SIZE || centroid.y < 0 || centroid.y > Tile.SIZE))
                     return false;