+++ docs/Changelog.md
@@ -2,17 +2,34 @@
 
 ## New since 0.9.2
 
+- Map overzoom [#393](https://github.com/mapsforge/vtm/issues/393)
+- Buildings overzoom [#503](https://github.com/mapsforge/vtm/issues/503)
 - S3DB layer [#475](https://github.com/mapsforge/vtm/pull/475)
 - vtm-mvt module with MVT tile decoder [#481](https://github.com/mapsforge/vtm/pull/481)
+- Nextzen MVT / GeoJSON vector tiles [#498](https://github.com/mapsforge/vtm/issues/498)
 - OpenMapTiles MVT vector tiles [#482](https://github.com/mapsforge/vtm/issues/482)
-- Theme styles improvements [#479](https://github.com/mapsforge/vtm/pull/479)
+- Location texture renderer [#547](https://github.com/mapsforge/vtm/issues/547)
+- Render themes: symbols on lines [#495](https://github.com/mapsforge/vtm/issues/495)
+- Render themes: styles improvements [#479](https://github.com/mapsforge/vtm/pull/479)
 - Internal render themes improvements [#488](https://github.com/mapsforge/vtm/pull/488)
 - Map view roll [#474](https://github.com/mapsforge/vtm/pull/474)
-- Fling animation improvements [#489](https://github.com/mapsforge/vtm/pull/489)
+- Physical fling and fling on rotation / scale [#497](https://github.com/mapsforge/vtm/pull/497) [#499](https://github.com/mapsforge/vtm/pull/499)
+  - `Parameters.ANIMATOR2 = true;`
+- Scale factor for short vertices calculation [#537](https://github.com/mapsforge/vtm/issues/537)
+  - `Parameters.CUSTOM_COORD_SCALE = false;`
+- Polygon symbols default disabled [#405](https://github.com/mapsforge/vtm/issues/405)
+  - `Parameters.POLY_SYMBOL = true;`
 - Map fractional zoom [#487](https://github.com/mapsforge/vtm/issues/487)
 - Render theme fallback internal resources [#477](https://github.com/mapsforge/vtm/issues/477)
+- Fix ItemizedLayer synchronization [#507](https://github.com/mapsforge/vtm/issues/507)
+- Fix Mapsforge clipping on small zooms [#264](https://github.com/mapsforge/vtm/issues/264)
+- Fix PathLayer (vtm) reverse segments [#220](https://github.com/mapsforge/vtm/issues/220)
 - Fix FadeStep alpha interpolation [#486](https://github.com/mapsforge/vtm/issues/486)
 - Fix libGDX flickering [#148](https://github.com/mapsforge/vtm/issues/148) [#149](https://github.com/mapsforge/vtm/issues/149)
+- ViewController refactor [#523](https://github.com/mapsforge/vtm/pull/523)
+  - `ViewController.setScreenSize` renamed to `setViewSize`
+  - `ViewController.setMapScreenCenter` renamed to `setMapViewCenter`
+- Android 9 compatibility [#550](https://github.com/mapsforge/vtm/issues/550)
 - JTS (LocationTech) [#484](https://github.com/mapsforge/vtm/issues/484)
 - Many other minor improvements and bug fixes
 - [Solved issues](https://github.com/mapsforge/vtm/issues?q=is%3Aclosed+milestone%3A0.10.0)
@@ -42,13 +59,14 @@
 - POI Search example [#394](https://github.com/mapsforge/vtm/issues/394)
 - Mapsforge Reverse Geocoding [#383](https://github.com/mapsforge/vtm/issues/383)
 - Core utilities [#396](https://github.com/mapsforge/vtm/issues/396)
-  - MercatorProjection.groundResolution renamed to groundResolutionWithScale
+  - `MercatorProjection.groundResolution` renamed to `groundResolutionWithScale`
 - Map scaling improvements [#401](https://github.com/mapsforge/vtm/issues/401)
 - Mapsforge fix artifacts zoom >17 [#231](https://github.com/mapsforge/vtm/issues/231)
 - Mapzen building extrusions [#419](https://github.com/mapsforge/vtm/issues/419)
   - BuildingLayer expects height tags in meters
 - Polygon symbol positioning [#405](https://github.com/mapsforge/vtm/issues/405)
 - PolyLabel default disabled [#402](https://github.com/mapsforge/vtm/issues/402)
+  - `Parameters.POLY_LABEL = true;`
 - vtm-theme-comparator module [#387](https://github.com/mapsforge/vtm/issues/387)
 - Feature parameters [#403](https://github.com/mapsforge/vtm/issues/403)
 - vtm-android-gdx module enhancements [#435](https://github.com/mapsforge/vtm/issues/435)
@@ -61,11 +79,13 @@
 ## Version 0.8.0 (2017-07-19)
 
 - Real time (SVG) texture atlas [#63](https://github.com/mapsforge/vtm/issues/63)
+  - `Parameters.TEXTURE_ATLAS = true;`
 - Marker clustering [#312](https://github.com/mapsforge/vtm/issues/312)
 - Osmagray theme [#300](https://github.com/mapsforge/vtm/issues/300)
 - Symbol rotation [#294](https://github.com/mapsforge/vtm/issues/294)
 - Location renderer improvements [#317](https://github.com/mapsforge/vtm/issues/317)
 - POT textures [#334](https://github.com/mapsforge/vtm/issues/334)
+  - `Parameters.POT_TEXTURES = true;`
 - OkHttp external cache [#135](https://github.com/mapsforge/vtm/issues/135)
 - Texture atlas improvements [#301](https://github.com/mapsforge/vtm/pull/301) [#304](https://github.com/mapsforge/vtm/pull/304)
 - vtm-json module [#367](https://github.com/mapsforge/vtm/issues/367)
@@ -85,7 +105,8 @@
 ## Version 0.7.0 (2017-02-26)
 
 - Mapsforge multiple map files [#208](https://github.com/mapsforge/vtm/issues/208)
-- Improved gestures implementation [#253](https://github.com/mapsforge/vtm/issues/253)
+- New gestures implementation [#253](https://github.com/mapsforge/vtm/issues/253)
+  - `Parameters.MAP_EVENT_LAYER2 = true;`
 - Polygon label position enhancements [#80](https://github.com/mapsforge/vtm/issues/80)
 - vtm-web modules update [#51](https://github.com/mapsforge/vtm/issues/51)
 - Mapzen MVT vector tiles [#57](https://github.com/mapsforge/vtm/issues/57)
@@ -121,6 +142,7 @@
 - Location renderer and layer [#171](https://github.com/mapsforge/vtm/issues/171)
 - Map scale bar [#84](https://github.com/mapsforge/vtm/issues/84)
 - Tile size based on scale factor [#183](https://github.com/mapsforge/vtm/issues/183)
+  - `Parameters.CUSTOM_TILE_SIZE = false;`
 - libGDX layer gestures [#151](https://github.com/mapsforge/vtm/issues/151)
 - Render theme area tessellation option [#37](https://github.com/mapsforge/vtm/issues/37)
 - Render theme resources optional location prefixes [#66](https://github.com/mapsforge/vtm/issues/66)